Backend. Fix some endpoints

+++ darkside-backend/darkside-backend/Controllers/ProductController.cs
@@ -1,5 +1,7 @@
 ﻿using darkside_backend.Helpers;
 using darkside_backend.Models.ApiModels;
+using darkside_backend.Models.Entities;
+using darkside_backend.Models.Enums;
 using darkside_backend.Services.Abstractions;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
@@ -17,10 +19,39 @@ namespace darkside_backend.Controllers
             _productService = productService ?? throw new ArgumentNullException(nameof(productService));
         }
 
+        [HttpGet]
+        public async Task<IActionResult> GetProduct([FromQuery] string id)
+        {
+            var response = await _productService.GetProductByIdAsync(id);
+
+            if (response == null)
+                return BadRequest(new { message = "Failed to fetch product" });
+
+            return Ok(response);
+        }
+
+        [HttpPost("list")]
+        public async Task<IActionResult> GetProductsByIds([FromBody] string[] ids)
+        {
+            var response = await _productService.GetProductsByIdsAsync(ids);
+
+            if (response == null)
+                return BadRequest(new { message = "Failed to get products" });
+
+            return Ok(response);
+        }
+
         [HttpPost]
         [Authorize]
         public async Task<IActionResult> CreateProduct([FromForm] ProductCreateRequest model)
         {
+            var user = (UserModel)HttpContext.Items["User"]!;
+
+            if (user.Role != RolesEnum.admin)
+            {
+                return Forbid("Only for admins!");
+            }
+
             var response = await _productService.CreateProductAsync(model);
 
             if (response == null)
@@ -33,6 +64,19 @@ namespace darkside_backend.Controllers
         [Authorize]
         public async Task<IActionResult> UpdateProduct(ProductUpdateRequest model)
         {
+            var user = (UserModel)HttpContext.Items["User"]!;
+
+            if (user.Role != RolesEnum.admin)
+            {
+                return Forbid("Only for admins!");
+            }
+
+            await _productService.UpdateTastes(new UpdateTastesRequest()
+            {
+                ProductId = model.Id,
+                TastesIds = model.TastesIds
+            });
+
             var response = await _productService.UpdateProductAsync(model);
             return Ok(response);
         }
@@ -41,6 +85,13 @@ namespace darkside_backend.Controllers
         [Authorize]
         public async Task<IActionResult> UpdateProductMainImage([FromForm] UpdateProductMainImageRequest model)
         {
+            var user = (UserModel)HttpContext.Items["User"]!;
+
+            if (user.Role != RolesEnum.admin)
+            {
+                return Forbid("Only for admins!");
+            }
+
             var response = await _productService.UpdateProductMainImageAsync(model);
             return Ok(response);
         }
@@ -49,24 +100,49 @@ namespace darkside_backend.Controllers
         [Authorize]
         public async Task<IActionResult> DeleteProduct([FromQuery] Guid id)
         {
+            var user = (UserModel)HttpContext.Items["User"]!;
+
+            if (user.Role != RolesEnum.admin)
+            {
+                return Forbid("Only for admins!");
+            }
+
             await _productService.DeleteProductAsync(id);
             return Ok();
         }
 
         [HttpGet("categories")]
-        [Authorize]
         public async Task<IActionResult> GetCategories()
         {
             var response = await _productService.GetCategories();
-
             return Ok(response);
         }
 
-        [HttpPost("search")]
-        [Authorize]
-        public async Task<IActionResult> GetCategories([FromQuery] string? term, [BindRequired, FromQuery] int limit, [BindRequired, FromQuery] int offset)
+        [HttpGet("creators")]
+        public async Task<IActionResult> GetCreators()
         {
-            var response = await _productService.SearchProducts(term, limit, offset);
+            var response = await _productService.GetCreators();
+            return Ok(response);
+        }
+
+        [HttpGet("search")]
+        public async Task<IActionResult> Search(
+            [BindRequired, FromQuery] int limit,
+            [BindRequired, FromQuery] int offset,
+            [FromQuery] string? term,
+            [FromQuery] string? category,
+            [FromQuery] string? creator,
+            [FromQuery] int priceMin = 0,
+            [FromQuery] int priceMax = int.MaxValue,
+            [FromQuery] int vpMin = 0,
+            [FromQuery] int vpMax = int.MaxValue
+            )
+        {
+            var user = (UserModel)HttpContext.Items["User"];
+            var response = await _productService.SearchProducts(term, limit, offset, category, creator, priceMin, priceMax, vpMin, vpMax, user is
+            {
+                Role: RolesEnum.admin
+            }, user is { Role: RolesEnum.admin });
 
             return Ok(response);
         }
@@ -75,6 +151,13 @@ namespace darkside_backend.Controllers
         [Authorize]
         public async Task<IActionResult> UpdateTastes(UpdateTastesRequest model)
         {
+
+            var user = (UserModel)HttpContext.Items["User"]!;
+
+            if (user.Role != RolesEnum.admin)
+            {
+                return Forbid("Only for admins!");
+            }
             await _productService.UpdateTastes(model);
 
             return Ok();
